+++ service/controller/cluster_resource_set.go
@@ -31,6 +31,7 @@ import (
 	"github.com/giantswarm/cluster-operator/service/controller/resource/cleanupmachinedeployments"
 	"github.com/giantswarm/cluster-operator/service/controller/resource/clusterconfigmap"
 	"github.com/giantswarm/cluster-operator/service/controller/resource/clusterid"
+	"github.com/giantswarm/cluster-operator/service/controller/resource/clusterstatus"
 	"github.com/giantswarm/cluster-operator/service/controller/resource/cpnamespace"
 	"github.com/giantswarm/cluster-operator/service/controller/resource/encryptionkey"
 	"github.com/giantswarm/cluster-operator/service/controller/resource/keepforinfrarefs"
@@ -217,17 +218,16 @@ func newClusterResourceSet(config clusterResourceSetConfig) (*controller.Resourc
 		}
 	}
 
-	var statusConditionResource resource.Interface
+	var clusterStatusResource resource.Interface
 	{
-		c := statuscondition.Config{
+		c := clusterstatus.Config{
 			K8sClient: config.K8sClient,
 			Logger:    config.Logger,
 
 			NewCommonClusterObjectFunc: config.NewCommonClusterObjectFunc,
-			Provider:                   config.Provider,
 		}
 
-		statusConditionResource, err = statuscondition.New(c)
+		clusterStatusResource, err = clusterstatus.New(c)
 		if err != nil {
 			return nil, microerror.Mask(err)
 		}
@@ -366,6 +366,22 @@ func newClusterResourceSet(config clusterResourceSetConfig) (*controller.Resourc
 		}
 	}
 
+	var statusConditionResource resource.Interface
+	{
+		c := statuscondition.Config{
+			K8sClient: config.K8sClient,
+			Logger:    config.Logger,
+
+			NewCommonClusterObjectFunc: config.NewCommonClusterObjectFunc,
+			Provider:                   config.Provider,
+		}
+
+		statusConditionResource, err = statuscondition.New(c)
+		if err != nil {
+			return nil, microerror.Mask(err)
+		}
+	}
+
 	var tenantClientsResource resource.Interface
 	{
 		c := tenantclients.Config{
@@ -456,6 +472,7 @@ func newClusterResourceSet(config clusterResourceSetConfig) (*controller.Resourc
 
 		// Following resources manage CR status information.
 		clusterIDResource,
+		clusterStatusResource,
 		statusConditionResource,
 
 		// Following resources manage tenant cluster deletion events.
@@ -484,8 +501,7 @@ func newClusterResourceSet(config clusterResourceSetConfig) (*controller.Resourc
 	}
 
 	initCtxFunc := func(ctx context.Context, obj interface{}) (context.Context, error) {
-		ctx = controllercontext.NewContext(ctx, controllercontext.Context{})
-		return ctx, nil
+		return controllercontext.NewContext(ctx, controllercontext.Context{}), nil
 	}
 
 	handlesFunc := func(obj interface{}) bool {