+++ service/controller/machine_deployment_resource_set.go
@@ -179,8 +179,7 @@ func newMachineDeploymentResourceSet(config machineDeploymentResourceSetConfig) 
 	}
 
 	initCtxFunc := func(ctx context.Context, obj interface{}) (context.Context, error) {
-		ctx = controllercontext.NewContext(ctx, controllercontext.Context{})
-		return ctx, nil
+		return controllercontext.NewContext(ctx, controllercontext.Context{}), nil
 	}
 
 	handlesFunc := func(obj interface{}) bool {